% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterAndSummarise.R
\name{calc.prev}
\alias{calc.prev}
\title{Calculate feature prevalence.}
\usage{
calc.prev(physeq)
}
\arguments{
\item{physeq}{(Required) A phyloseq object.}
}
\value{
A data frame.
}
\description{
Function to create a data frame showing the prevalence of each feature across all samples plus extra infromation.
}
\examples{
df <- calc.prev(phyloseq.object)

}
