% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{clr_runif}
\alias{clr_runif}
\title{CLR (centered log-ratio) transformation of phyloseq object (pseudocount equal to random numbers between 0.1*min_value and min_value).}
\usage{
clr_runif(physeq)
}
\arguments{
\item{physeq}{(Required) A phyloseq object.}
}
\value{
A phyloseq object with transformed abundance data.
}
\description{
This function performs the Chord transformation on the abundance data of a phyloseq object.
The pseudocounts used to avoid 0s when log transforming are random numbers between the minimum abundance value (min_value) and 0.1*min_value.
This transformation should be used with raw abundance data (internal Total Sum Scaling transformation -- relative abundances).
}
\examples{
data("GlobalPatterns")
transformed.data <- clr_runif(GlobalPatterns)

}
\references{
Lubbe S. et al.  (2021) Comparison of zero replacement strategies for compositional data with large numbers of zeros. Chemometr. Intell. Lab., 210, 104248.
}
