% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterAndSummarise.R
\name{FilterAndSummarise}
\alias{FilterAndSummarise}
\title{Filter and produce summary files}
\usage{
FilterAndSummarise(
  physeq,
  file.dir = NULL,
  filt.domain = c("Unclassified", "Eukaryota", "Viruses", "Archaea"),
  min.depth = 1000,
  min.prev = 2,
  min.abund = 10,
  abund.type = "Total",
  report = FALSE,
  seqrows = FALSE
)
}
\arguments{
\item{physeq}{(Required) A phyloseq object.}

\item{file.dir}{(Optional) A character string of the PATH defining the directory where the summary files and plots are going to be created.
If no directory is specified files won't be created. DEFAULT: NULL}

\item{filt.domain}{(Optional) A character string defining the names of the domains to filter out. DEFAULT: c("Unclassified", "Eukaryota", "Viruses", "Archaea").}

\item{min.depth}{(Optional) A numerical value -- a threshold -- indicating the minimum number (>=) of reads a sample should have. DEFAULT: 1000}

\item{min.prev}{(Optional) A numerical value -- a threshold -- indicating the minimum number (>=) of samples where each taxon should be present. DEFAULT: 2}

\item{min.abund}{(Optional) A numerical value -- a threshold -- indicating the minimum abundance (>=) a taxon should have. DEFAULT: 10}

\item{abund.type}{(Optional) A character string defining which type of abundance should be taken into account when filtering ("Total" or "Mean") DEFAULT: "Mean"}

\item{report}{(Optional) A logical value. If TRUE an excel summary file is created. DEFAULT: FALSE}

\item{seqrows}{(Optional) A numerical value indicating if rownames are amplicon DNA sequences. If TRUE, a file showing the distribution of sequence lengths between filtered and unfiltered taxa is created. DEFAULT: FALSE.}
}
\value{
A filtered phyloseq object. It also writes summary files in the specified directory.
}
\description{
This function is used to filter samples with insufficient read counts, undesired taxa such as order Chloroplast and family Mitochondria and low prevalence taxa.
Furthermore, all Phyla with not available/missing taxonomic classification are removed.
Filtration at the domain rank can be customised.
}
\examples{
data("GlobalPatterns")
colnames(tax_table(GlobalPatterns)) <- c("Domain", "Phylum", "Class", "Order", "Family", "Genus", "Species")
To filter out order Chloroplast, family Mitochondria and domains Eukaryota, Archaea, Viruses and Unclassified:
data_phylo <- FilterAndSummarise(GlobalPatterns, min.depth = 0, min.prev = 0, min.abund = 0)

To perform the previous filtration but also obtain the summary files and carry out an abundance filtration based on total abundance:
data_phylo <- FilterAndSummarise(GlobalPatterns, "~/output/summary_files", min.depth = 1000, min.prev = 3, min.abund = 20, abund.type = "Total")

}
