% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{hellinger}
\alias{hellinger}
\title{Hellinger transformation of phyloseq object.}
\usage{
hellinger(physeq)
}
\arguments{
\item{physeq}{(Required) A phyloseq object.}
}
\value{
A phyloseq object with transformed abundance data.
}
\description{
This function performs the Hellinger transformation on the abundance data of a phyloseq object.
This transformation should be used with raw abundance data.
}
\examples{
data("GlobalPatterns")
transformed.data <- hellinger(GlobalPatterns)

}
