% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{logchord_1p}
\alias{logchord_1p}
\title{Chord transformation of phyloseq object (pseudocount of 1 added to all values).}
\usage{
logchord_1p(physeq)
}
\arguments{
\item{physeq}{(Required) A phyloseq object.}
}
\value{
A phyloseq object with transformed abundance data.
}
\description{
This function performs the Chord transformation on the abundance data of a phyloseq object.
The pseudocount used to avoid 0s when log transforming is 1.
This transformation should be used with raw abundance data.
}
\examples{
data("GlobalPatterns")
transformed.data <- logchord_1p(GlobalPatterns)

}
