% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D_plots.R
\name{MDSin3D}
\alias{MDSin3D}
\title{3D plots}
\usage{
MDSin3D(
  physeq,
  explanatory_variable,
  distance = "bray",
  method = "PCoA",
  colors,
  constrained = FALSE,
  formula
)
}
\arguments{
\item{physeq}{(Required) A phyloseq object.}

\item{explanatory_variable}{(Required) A character string of the metadata variable that is going to be used to group the samples.}

\item{distance}{A character string defining the distance used to calculate the distance matrix:
It can be any distance accepted by the distance function of the phyloseq package and also "PhilR".
To use the Aitchison distance, first transform the otu_table of the phyloseq object with one of the CLR functions
(clr_min_percent, clr_0_plus_1 or clr_runif) and then use the "euclidean" distance.
Default: "bray"}

\item{method}{A character string defining the ordination method: "PCA", "PCoA", "RDA", "CAP", "CA", "CCA" or "NMDS"
Default: "PCoA"}

\item{colors}{(Optional) A vector with the hexadecimal colors associated to the levels of the explanatory variable.}

\item{constrained}{A logical value. If TRUE a constrained ordination is carried out.
Default: FALSE}

\item{formula}{(Optional) The right hand side of the model formula -- the constraining variables and conditioning variables. e.g: "~ constraining variable"
Only used if constrained = TRUE}
}
\value{
A plotly object.
}
\description{
This function is used to create a MultiDimensional Scaling ordination plot in 3D with plotly.
}
\examples{
Unconstrained ordination plot:
plot_3D <- MDSin3D(physeq = phyloseq.object, explanatory_variable = "Sex", colors = c("#F8766D", "#00C064"))

Constrained ordination plot:
plot_3D <- MDSin3D(physeq = phyloseq.object, explanatory_variable = "Sex", constrained = TRUE, formula = "~ Country")
}
