% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{rare.curve}
\alias{rare.curve}
\title{Rarefaction Curves}
\usage{
rare.curve(
  physeq,
  vegandata = NULL,
  step = 500,
  col_var = NULL,
  line_var = NULL,
  colors = NULL,
  linetypes = NULL,
  labels = TRUE,
  title = NULL,
  y_lab = NULL,
  legend.text.size = 1,
  file_name = "plot",
  legend.box.col = c(0.9, 0.97),
  legend.box.line = c(0.9, 0.6),
  legend.box1.width = 0.3,
  legend.box2.width = 0.3,
  legend.box1.height = 0,
  legend.box2.height = 0,
  legend.titlecol = "#00008B",
  legend.back = "#D3D3D3",
  str2rm = "",
  return.vegandata = TRUE,
  return.plot = TRUE
)
}
\arguments{
\item{physeq}{(Required) A phyloseq object.}

\item{vegandata}{(Optional) Output from the rarecurve function of the Vegan package.
Providing this data reduces the computational time, therefore changing graphical parameters is quickly performed. DEFAULT: NULL}

\item{step}{(Required) Step size for sample sizes in rarefaction curves. DEFAULT: 500}

\item{col_var}{(Optional) A character string indicating the metadata variable used to set the color of the lines. DEFAULT: NULL}

\item{line_var}{(Optional) A character string indicating the metadata variable used to set the linetype of the lines. DEFAULT: NULL}

\item{colors}{(Optional) A vector with the hexadecimal colors associated to the levels of the color variable. DEFAULT: NULL}

\item{linetypes}{(Optional) A vector with the line type names associated to the levels of the linetype variable. DEFAULT: NULL}

\item{labels}{(Optional) A logical value. If TRUE sample names are shown at the end of each curve. DEFAULT: TRUE}

\item{title}{(Optional) A character string indicating the plot title. DEFAULT: NULL}

\item{file_name}{(Required) A character string indicating the name of the png plot file. DEFAULT: "plot"}

\item{legend.box.col}{(Optional) A vector indicating the coordinates (position) of the color legend box. DEFAULT: c(0.86, 0.9)}

\item{legend.box.line}{(Optional) A vector indicating the coordinates (position) of the linetype legend box. DEFAULT: c(0.825, 0.4)}

\item{legend.box1.width}{(Optional) A numerical value indicating the width of the legend box 1. DEFAULT: 0.3}

\item{legend.box2.width}{(Optional) A numerical value indicating the width of the legend box 2. DEFAULT: 0.3}

\item{legend.box1.height}{(Optional) A numerical value indicating the height of the legend box 1. DEFAULT: 0}

\item{legend.box2.height}{(Optional) A numerical value indicating the height of the legend box 2. DEFAULT: 0}

\item{legend.titlecol}{(Optional) A character string indicating the hexadecimal color used for the legend title. DEFAULT: "#00008B"}

\item{legend.back}{(Optional) A character string indicating the hexadecimal color used for the legend background. DEFAULT: "#D3D3D3"}

\item{str2rm}{(Optional) A vector of character strings indicating the patterns to be removed from the names of the samples. DEFAULT: ""}

\item{return.vegandata}{(Optional) A logical value. If TRUE a plot_grid object is returned. If FALSE plot is saved in the specified directory. DEFAULT: TRUE}

\item{return.plot}{(Optional) A logical value. If TRUE the output from the rarecurve function of the Vegan package is returned. DEFAULT: FALSE}

\item{ylab}{(Optional) A character string indicating the title of the y axis. DEFAULT: "Number of different ASVs"}
}
\value{
A filtered phyloseq object. It also writes summary files in the specified directory.
Or if out.table = TRUE a data frame showing the prevalence of each feature (OTU/ASV/Species) across all samples.
}
\description{
This function is used to create a plot that shows the rarefaction curves of all samples.
Curves can be colored or have different linetypes depending on the grouping variables specified.
}
\examples{
data("GlobalPatterns")
setwd("~/PATHtoDirectory")
rare.curve(GlobalPatterns, file_name = "Rarefaction_curves", col_var = "SampleType")

}
