% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Taxonomic.analyis}
\alias{Taxonomic.analyis}
\title{Taxonomic profile, count data and sample clustering tree plots.}
\usage{
Taxonomic.analyis(
  physeq,
  abundance.plot = "pointplot",
  count.plot = FALSE,
  data.count.plot = NULL,
  count.plot.rel = FALSE,
  domains = c("Archaea", "Bacteria", "Eukaryota", "Unclassified"),
  dendogram = FALSE,
  rank = "Phylum",
  ntaxa = 15,
  taxa.txt.size = 12,
  sample.txt.size = 8,
  show.unidentified = TRUE,
  name.format = "short",
  ASV.count = TRUE,
  col.palette = c("#ff0000", "#00ff00", "#0000ff"),
  title = "Bacterial phyla",
  return.plot = TRUE,
  folder = NULL,
  file.name = "plot.png",
  str2rm = NULL,
  Top.var = NULL,
  Top.strip.fill = "by.Top.var",
  Bot.var = NULL,
  Bot.strip.fill = "by.Bot.var",
  nboot = 500,
  aglom.clust = FALSE,
  transf = "CLR_runif",
  distance = "Euclidean",
  clust.method = "ward.D",
  n.clusters = 2,
  clust.data.rel.abund = FALSE,
  nudge.bs.label = NULL,
  bs.label.size = 2.5,
  icons.dendogram = FALSE,
  icons.variable = NULL,
  icons.folder = NULL,
  tile.col = "blank"
)
}
\arguments{
\item{physeq}{(Required) A phyloseq object.}

\item{abundance.plot}{(Optional) A character string indicating how to display the taxonomic profile ("barplot", "poinplot" or "blank"). DEFAULT: "pointplot"}

\item{count.plot}{(Optional) A logical value. If TRUE a stacked bar plot with count data per domain is displayed. DEFAULT: FALSE}

\item{data.count.plot}{(Optional, only required if count.plot = TRUE) A phyloseq object with the data used to create the count.plot.
Since this plot is made to provide library depth information unfiltered data should be provided. DEFAULT: NULL}

\item{count.plot.rel}{(Optional)}

\item{domains}{(Optional)}

\item{dendogram}{(Optional) A logical value. If TRUE a clustering tree diagram of the samples is displayed. DEFAULT: FALSE}

\item{rank}{(Optional) A character string indicating the taxonomic rank shown in the abundance.plot. DEFAULT: "Phylum"}

\item{ntaxa}{(Optional) A numerical value indicating the number of taxa shown in the abundance.plot. DEFAULT: 15}

\item{taxa.txt.size}{(Optional) A numerical value indicating the text size of the taxonomic names.
Only required if an abundance.plot is created. DEFAULT: 12}

\item{sample.txt.size}{(Optional) A numerical value indicating the text size of the sample names. DEFAULT: 8}

\item{show.unidentified}{(Optional) A logical value. If TRUE the unclassified/n.a taxa are renamed using the information their higher taxonomic ranks. DEFAULT: TRUE}

\item{name.format}{(Optional) A character string indicating the name format of the renamed unclassified taxa ("short" or "long", for more info: \link[metaGVis]{unclassified_ranks}). DEFAULT: "short"}

\item{ASV.count}{(Optional) A logical value. If TRUE the the number of species agglomerated under a higher taxonomic rank is appended at the end of the taxonomic name. DEFAULT: TRUE}

\item{col.palette}{(Optional) A character vector containing the hexadecimal representations of one or more colors (seedcolors). DEFAULT: c("#ff0000", "#00ff00", "#0000ff")}

\item{title}{(Optional) Main plot title. DEFAULT: ""}

\item{return.plot}{(Optional) A logical value. If TRUE a plot_grid object is returned. If FALSE plot is saved in the specified directory. DEFAULT: TRUE}

\item{folder}{(Optional) A character string indicating the PATH to the directory where the plot is saved. DEFAULT: NULL}

\item{file.name}{(Optional) File name. DEFAULT: "plot.png"}

\item{str2rm}{(Optional) A vector of character strings defining the patterns to be removed from the names of the samples. DEFAULT: NULL}

\item{Top.var}{(Optional) A character string indicating the data frame variable used in facet_wrap.
For more info: \link[metaGVis]{wrap.strip.colors}). DEFAULT: NULL}

\item{Top.strip.fill}{(Optional) A vector with the hexadecimal colors associated to the levels of the Top variable.
OR a character string ("by.Top.var" or "by.Bot.var" or "blank") indicating if colors are assigned to the levels of the Top or Bottom variables.
For more info: \link[metaGVis]{wrap.strip.colors}). DEFAULT: "by.Top.var"}

\item{Bot.var}{(Optional) A character string indicating a second data frame variable used in facet_wrap. DEFAULT: NULL}

\item{Bot.strip.fill}{(Optional) A vector with the hexadecimal colors associated to the levels of the Bottom variable.
OR a character string ("by.Top.var", "by.Bot.var" or "blank") indicating if colors are assigned to the levels of the Top or Bottom variables. DEFAULT: "by.Bot.var"}

\item{nboot}{(Optional) A numeric value indicating the number of bootstrap replications. DEFAULT: 500}

\item{aglom.clust}{(Optional) A logical value. If TRUE the clustering tree diagram of the samples is performed using the agglomerated data at the specified rank. DEFAULT: FALSE}

\item{transf}{(Optional) A character string indicating the transformation performed on the data before clustering
(Hellinger, logChord_0plus1, logChord_1p, CLR_0plus1, CLR_min or CLR_runif). DEFAULT: "CLR_runif"}

\item{distance}{(Optional) A character string indicating the clustering distance (Bray, wUniFrac, UniFrac, Euclidean). DEFAULT: "Euclidean"}

\item{clust.method}{(Optional) A character string indicating the hierarchical clustering algorithm used (single, complete, average, ward.D). DEFAULT: "ward.D"}

\item{n.clusters}{(Optional) A numeric value indicating the number of clusters to color code. DEFAULT: 2}

\item{clust.data.rel.abund}{(Optional) A logical value. If TRUE the abundance data is transformed into relative abundance before clustering (only applicable for Bray and wUniFrac). DEFAULT: FALSE}

\item{nudge.bs.label}{(Optional) A numeric value (see y.axis range) indicating how much the bootstrap values are moved vertically. DEFAULT: NULL}

\item{bs.label.size}{(Optional) A numeric value indicating the size of the bootstrap values. DEFAULT: 2.5}

\item{icons.dendogram}{(Optional) A logical value. If TRUE specified icons representing samples are located at the end of the dendogram branches. DEFAULT: FALSE}

\item{icons.variable}{(Optional) A character string indicating the variable use to assign icons. DEFAULT: NULL}

\item{icons.folder}{(Optional) A character string indicating the PATH to the directory where the icon files are located.
PNG files must have the same names as the icons.variable levels. DEFAULT: NULL}

\item{tile.col}{(Optional) A vector with the hexadecimal colors associated to the levels of the Bottom variable. The selected variable is used to color code the tile found at the end of the dendogram branches.
OR a character string ("by.Top.var", "by.Bot.var" or "blank") indicating if colors are assigned to the levels of the Top or Bottom variables. DEFAULT: "by.Bot.var"}
}
\value{
The plot is saved as a png file in the specified directory. If return.plot = TRUE, it returns a plot_grid() object.
}
\description{
This function can create 3 different plots: a taxonomic profile (barplot or poinplot), a stacked bar plot with count data per domain
and a clustering tree diagram of the samples. The 3 plots can be visualised individually or together.
}
\examples{
data("GlobalPatterns")
# Since all metaGVis functions use "Domain" rank instead of Kingdom, let's change it:
colnames(tax_table(GlobalPatterns)) <- c("Domain", "Phylum", "Class", "Order", "Family", "Genus", "Species")
# To make sure the first metadata column is named "SampleID":
colnames(sample_data(GlobalPatterns))[1] <- "SampleID"
filt.data <- FilterAndSummarise(GlobalPatterns)

stackedbar.plot <- Taxonomic.analyis(filt.data, abundance.plot = "barplot", rank = "Order",
title = "Relative abundance of the 15 most predominant bacterial orders")
print(stackedbar.plot)

stackedbar.plot.with.facets.and.1.variable <- Taxonomic.analyis(filt.data, abundance.plot = "barplot", rank = "Order",
Top.var = "SampleType", title = "Relative abundance of the 15 most predominant bacterial orders")
print(stackedbar.plot.with.facets.and.1.variable)

stackedbar.plot.with.facets.and.2.variables <- Taxonomic.analyis(filt.data, abundance.plot = "barplot", rank = "Order",
Top.var = "SampleType", Top.strip.fill = "by.Bot.var", Bot.var = "Primer", Bot.strip.fill = "by.Bot.var",
title = "Relative abundance of the 15 most predominant bacterial orders")
print(stackedbar.plot.with.facets.and.2.variables)

point.plot <- Taxonomic.analyis(filt.data, abundance.plot = "pointplot", rank = "Order",
title = "Relative abundance of the 15 most predominant bacterial orders")
print(point.plot)

count.plot <- Taxonomic.analyis(data.count.plot = GlobalPatterns, count.plot = TRUE, abundance.plot = "blank",
title = "Relative abundance of the 15 most predominant bacterial orders")
print(count.plot)

combined.pointplot <- Taxonomic.analyis(filt.data, data.count.plot = GlobalPatterns, abundance.plot = "pointplot", rank = "Order", count.plot = TRUE, dendogram = TRUE, nboot = 1,
title = "Relative abundance of the 15 most predominant bacterial orders")
print(combined.plot)

combined.barplot.with.1.variable <- Taxonomic.analyis(filt.data, data.count.plot = GlobalPatterns, abundance.plot = "barplot", rank = "Phylum",
count.plot = TRUE, dendogram = T, nboot = 1, Top.var = "SampleType", tile.col = "by.Top.var",
title = "Relative abundance of the 15 most predominant bacterial orders")
print(combined.barplot.with.variables)

combined.pointplot.with.1.variable <- Taxonomic.analyis(filt.data, data.count.plot = GlobalPatterns, abundance.plot = "pointplot", rank = "Phylum",
count.plot = TRUE, dendogram = T, nboot = 1, Top.var = "SampleType", tile.col = "by.Top.var",
title = "Relative abundance of the 15 most predominant bacterial orders")

combined.pointplot.with.2.variables <- Taxonomic.analyis(filt.data, data.count.plot = GlobalPatterns, abundance.plot = "pointplot", rank = "Phylum",
count.plot = TRUE, dendogram = F, nboot = 1, Top.var = "SampleType", Bot.var = "Primer", Top.strip.fill = "blank", Bot.strip.fill = "by.Bot.var",
title = "Relative abundance of the 15 most predominant bacterial orders")

}
